# Rotation calculated for spherical particles