--- a/cmake/modules/ALL.cmake
+++ b/cmake/modules/ALL.cmake
@@ -1,10 +1,5 @@
 # cmake module for adding ALL
 
-option(ENABLE_GPROF "Use the GNU profiler (gprof)" OFF)
-if(ENABLE_GPROF)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -pg")
-endif()
-
 option(ENABLE_ALLLBL "Enable ALL load balancing library" OFF)
 if(ENABLE_ALLLBL)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DENABLE_ALLLBL")