… doc/profiling.dox