# Description
Optimize rotation of zero vectors (spherical molecule simulations).  
This case is handled explicitly to reduce the number of multiplications that are performed.
$\forall q\in\mathbb{H}:rotate(q,\vec{\scriptstyle 0})=\vec{\scriptstyle 0}$

For `Quaternion::differentiate` all summands in all equations have a factor from the input vector.
Thus, $\forall q\in\mathbb{H}:differentiate(q,\vec{\scriptstyle 0})=\vec{\scriptstyle 0}$

## Resolved Issues

- [ ] #267 

## How Has This Been Tested?
As described in #267, the Argon and CO2 simulations were run using both the original and modified codes generating .xyz-files.
There was no difference between the .xyz-files which indicates that output of the computations is unchanged.