@rubenhorn 2023-09-18
Other optimizations may also be implemented :point_right: Do NOT merge yet!