@rubenhorn 2023-11-07
In a very naive sequential test using C++/gcc, avoiding 100M calls to `sqrt(1.0)` for saves 10ms.
This is not much, but it may be notable when running very large simulations (particle count/time steps) without AutoPas.