@rubenhorn 2023-11-07
Using the GNU profiler (gprof) for the argon example (`config_noAP_ALL.xml` with `<steps>10000</steps>`, no output):
| Method | baseline (103.45s)  | optimized (95.71s)  |
|---|---|---|
| `Quaternion::rotate` | 0.56s (0.79%) | 0.19s (0.30%) |
| `Quaternion::rotateinv` | 0.21s (0.30%) | 0.08s (0.12%) |
| `Quaternion::differentiate` | 0.06s (0.09%) | 0.10s (0.16%) |

Reverting the change of `Quaternion::differentiate` .